requirements.txt
beautifulsoup4==4.10.0
Django==4.0.2
djangorestframework==3.14.0
Babel==2.10.3
requests==2.26.0
requests-oauthlib==1.3.0
Pillow==9.2.0
psycopg2-binary
celery==5.2.7
redis==4.5.4
django-celery-beat